Update i3 config with improvements inspired from https://github.com/bookercodes/dotfiles

+++ install.sh
@@ -7,8 +7,8 @@ popd > /dev/null
 function install_file {
   SRC=$1
   DST=$2
-  [[ ! -s $DST ]] && mv $DST $DST.old
-  ln -sf $SRC $DST
+  if [[ -f $DST && ! -s $DST ]] ; then mv $DST $DST.old; fi
+  if [[ ! -s $DST ]] ; then ln -sf $SRC $DST; fi
 }
 install_file "$SCRIPTPATH/i3blocks.conf" ~/.i3blocks.conf
 install_file "$SCRIPTPATH/.bashrc" ~/.bashrc
@@ -16,7 +16,14 @@ install_file "$SCRIPTPATH/.prompt" ~/.prompt
 install_file "$SCRIPTPATH/.wallpaper.jpg" ~/.wallpaper.jpg
 install_file "$SCRIPTPATH/.bash_aliases" ~/.bash_aliases
 
+mkdir -p ~/.fonts
+cp $SCRIPTPATH/.fonts/* ~/.fonts/
+
 mkdir -p ~/.config/i3/
 install_file "$SCRIPTPATH/.config/i3/config" ~/.config/i3/config
-i3-msg restart
 
+sudo apt-get update
+sudo apt-get -y upgrade
+sudo apt-get -y install terminator i3 i3blocks redshift-gtk keepass2 pass undistract-me steam python3-dev python3-pip rofi
+sudo pip3 install virtualenvwrapper
+sudo dpkg-reconfigure keyboard-configuration
